(#68) mundando as pastas para a arquitetura factory+blueprints
