--- src/app/controller/__init__.py
@@ -1,4 +1,0 @@
-def configure(app):
-    pass
-
-
