--- src/app/model/__init__.py
@@ -1,7 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
-
-def configure(app):
-    db.init_app(app)
-    app.db = db
