--- src/app/view/__init__.py
@@ -1,6 +1,0 @@
-from flask_marshmallow import Marshmallow
-
-ma = Marshmallow()
-
-def configure(app):
-    ma.init_app(app)
